+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.15.0 (2020-06-07)
+-------------------
+
+* Switch pml fixedpoint iteration thresholds to the Margalit--Strenner--Yurtas bounds.
+* Added heuristics for finding monodromies of bundles using taut structures.
+  Joint with Nathan M Dunfield.
+
 0.14.0 (2020-02-18)
 -------------------